% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapOccur.R
\name{mapOccur}
\alias{mapOccur}
\title{Mapea los puntos de una tabla Ocurrencia}
\usage{
mapOccur(data = data, colId = "id", buffer = 0)
}
\arguments{
\item{data}{un dataframe con la tabla EventCore}

\item{colId}{el nombre de la columna con ID entre comillas}

\item{buffer}{valor numerico en kilometros para aplicar buffer a cada punto}
}
\value{
los puntos ubicados en el mapa
}
\description{
Genera un mapa dinámico con los puntos georeferenciados de una tabla EventCore.
Cada punto será de color verde si cae en el oceano o de color rojo si cae en tierra.
Al hacer clic en el punto aparece la información de eventID, localidad, y coordenadas.
}
\author{
Carlos Zamora-Manzur \email{carzamora@gmail.com}
}
